feat(size03): split website-db.ts into domain modules (4436→2890 lines) [T001293]
Extract four domain-specific modules from the god-file website-db.ts:
- appointments-db.ts: calendar/booking/scheduling domain (468 LOC)
- meetings-db.ts: meeting lifecycle, transcripts, artifacts, insights (365 LOC)
- billing-db.ts: billing schema init — initBillingTables, initTaxMonitorTables, initEurTables (618 LOC)
- test-infra-db.ts: test runs, test results, flake detection, playwright reports (251 LOC)

All exports preserved via re-export stubs in website-db.ts for full backward compatibility.
Removes website-db.ts from s1.ignore in gates.yaml — G-SIZE03 gate now green (2890 ≤ 3000).
