+++ website/src/lib/website-db.ts
@@ -10,6 +10,11 @@ import { refFor } from './content-registry';
 import { idsToPrune } from './admin/version-prune';
 import { isConflict as detectConflict, nextVersion as bumpVersion } from './admin/conflict';
 
+// Cross-domain imports for meeting-project linkage functions (Meeting-Projekt-Verknüpfung)
+// that remain in website-db.ts but need helpers from meetings-db.ts (G-SIZE03).
+import { initMeetingProjectLink } from './meetings-db';
+import type { MeetingWithDetails } from './meetings-db';
+
 // pool / ensureSchemaOnce / platformPool were moved to db-pool.ts (G-CQ07) so
 // tickets modules can depend on the leaf-most pool without re-entering
 // website-db. Re-export for backward compatibility with any external caller
@@ -247,359 +252,28 @@ export async function setIsAdmin(customerId: string, isAdmin: boolean): Promise<
   await pool.query('UPDATE customers SET is_admin = $1 WHERE id = $2', [isAdmin, customerId]);
 }
 
-// ── Schema init ─────────────────────────────────────────────────────────────
-
-export async function initMeetingsDb(): Promise<void> {
-  await pool.query(`
-    CREATE TABLE IF NOT EXISTS meetings (
-      id UUID PRIMARY KEY DEFAULT gen_random_uuid(),
-      customer_id UUID NOT NULL REFERENCES customers(id),
-      meeting_type TEXT NOT NULL,
-      scheduled_at TIMESTAMPTZ,
-      talk_room_token TEXT,
-      status TEXT NOT NULL DEFAULT 'scheduled',
-      started_at TIMESTAMPTZ,
-      ended_at TIMESTAMPTZ,
-      duration_seconds INTEGER,
-      recording_path TEXT,
-      released_at TIMESTAMPTZ,
-      created_at TIMESTAMPTZ NOT NULL DEFAULT now(),
-      updated_at TIMESTAMPTZ NOT NULL DEFAULT now()
-    )
-  `);
-  await pool.query(
-    'ALTER TABLE meetings ADD COLUMN IF NOT EXISTS released_at TIMESTAMPTZ'
-  );
-}
-
-async function initMeetingProjectLink(): Promise<void> {
-  await initTicketsSchema(); // tickets.tickets must exist before the FK column
-  await pool.query(`
-    ALTER TABLE meetings
-      ADD COLUMN IF NOT EXISTS project_id UUID REFERENCES tickets.tickets(id) ON DELETE SET NULL
-  `);
-  await pool.query(`
-    CREATE INDEX IF NOT EXISTS idx_meetings_project ON meetings(project_id)
-  `);
-}
-
-// ── Meeting ─────────────────────────────────────────────────────────────────
-
-export interface Meeting {
-  id: string;
-  customerId: string;
-  status: string;
-  meetingType: string;
-  scheduledAt: Date | null;
-  createdAt: Date;
-  released_at: Date | null;
-  projectId: string | null;
-  projectName: string | null;
-}
-
-export interface MeetingWithDetails {
-  id: string;
-  meetingType: string;
-  status: string;
-  scheduledAt: Date | null;
-  startedAt: Date | null;
-  endedAt: Date | null;
-  durationSeconds: number | null;
-  releasedAt: Date | null;
-  createdAt: Date;
-  transcripts: Array<{
-    id: string;
-    fullText: string;
-    language: string;
-    durationSeconds: number | null;
-  }>;
-  insights: Array<{
-    id: string;
-    insightType: string;
-    content: string;
-    generatedBy: string;
-  }>;
-  artifacts: Array<{
-    id: string;
-    artifactType: string;
-    name: string;
-    contentText: string | null;
-  }>;
-}
-
-export interface MeetingWithCustomer {
-  id: string;
-  customerId: string;
-  customerName: string;
-  customerEmail: string;
-  meetingType: string;
-  status: string;
-  talkRoomToken: string | null;
-}
-
-export async function getMeetingByRoomToken(
-  roomToken: string,
-): Promise<MeetingWithCustomer | null> {
-  const result = await pool.query(
-    `SELECT m.id, m.customer_id AS "customerId",
-            c.name AS "customerName", c.email AS "customerEmail",
-            m.meeting_type AS "meetingType", m.status,
-            m.talk_room_token AS "talkRoomToken"
-     FROM meetings m
-     JOIN customers c ON m.customer_id = c.id
-     WHERE m.talk_room_token = $1
-     ORDER BY m.created_at DESC
-     LIMIT 1`,
-    [roomToken],
-  );
-  return result.rows[0] ?? null;
-}
-
-export async function createMeeting(params: {
-  customerId: string;
-  meetingType: string;
-  scheduledAt?: Date;
-  talkRoomToken?: string;
-  projectId?: string;
-}): Promise<Meeting> {
-  const result = await pool.query(
-    `INSERT INTO meetings (customer_id, meeting_type, scheduled_at, talk_room_token, status, project_id)
-     VALUES ($1, $2, $3, $4, 'scheduled', $5)
-     RETURNING id, customer_id as "customerId", status, released_at,
-               project_id as "projectId", NULL::text as "projectName"`,
-    [params.customerId, params.meetingType, params.scheduledAt,
-     params.talkRoomToken, params.projectId ?? null]
-  );
-  return result.rows[0];
-}
-
-export async function updateMeetingStatus(meetingId: string, status: string, extra?: {
-  startedAt?: Date;
-  endedAt?: Date;
-  durationSeconds?: number;
-  recordingPath?: string;
-}): Promise<void> {
-  const sets = ['status = $2'];
-  const values: unknown[] = [meetingId, status];
-  let idx = 3;
-
-  if (extra?.startedAt) { sets.push(`started_at = $${idx}`); values.push(extra.startedAt); idx++; }
-  if (extra?.endedAt) { sets.push(`ended_at = $${idx}`); values.push(extra.endedAt); idx++; }
-  if (extra?.durationSeconds) { sets.push(`duration_seconds = $${idx}`); values.push(extra.durationSeconds); idx++; }
-  if (extra?.recordingPath) { sets.push(`recording_path = $${idx}`); values.push(extra.recordingPath); idx++; }
-
-  await pool.query(`UPDATE meetings SET ${sets.join(', ')} WHERE id = $1`, values);
-}
-
-// ── Transcript ──────────────────────────────────────────────────────────────
-
-export interface SavedTranscript {
-  id: string;
-  meetingId: string;
-}
-
-export async function saveTranscript(params: {
-  meetingId: string;
-  fullText: string;
-  language?: string;
-  whisperModel?: string;
-  durationSeconds?: number;
-  segments?: Array<{ start: number; end: number; text: string; speaker?: string }>;
-}): Promise<SavedTranscript> {
-  const client = await pool.connect();
-  try {
-    await client.query('BEGIN');
-
-    const transcriptResult = await client.query(
-      `INSERT INTO transcripts (meeting_id, full_text, language, whisper_model, duration_seconds)
-       VALUES ($1, $2, $3, $4, $5)
-       RETURNING id`,
-      [params.meetingId, params.fullText, params.language || 'de',
-       params.whisperModel || 'Systran/faster-whisper-medium', params.durationSeconds]
-    );
-    const transcriptId = transcriptResult.rows[0].id;
-
-    if (params.segments && params.segments.length > 0) {
-      for (let i = 0; i < params.segments.length; i++) {
-        const seg = params.segments[i];
-        await client.query(
-          `INSERT INTO transcript_segments (transcript_id, segment_index, start_time, end_time, text, speaker)
-           VALUES ($1, $2, $3, $4, $5, $6)`,
-          [transcriptId, i, seg.start, seg.end, seg.text, seg.speaker]
-        );
-      }
-    }
-
-    await client.query('COMMIT');
-    return { id: transcriptId, meetingId: params.meetingId };
-  } catch (err) {
-    await client.query('ROLLBACK');
-    throw err;
-  } finally {
-    client.release();
-  }
-}
-
-// ── Artifacts ───────────────────────────────────────────────────────────────
-
-export async function saveArtifact(params: {
-  meetingId: string;
-  artifactType: 'whiteboard' | 'document' | 'screenshot' | 'file';
-  name: string;
-  storagePath?: string;
-  contentText?: string;
-  metadata?: Record<string, unknown>;
-}): Promise<string> {
-  const result = await pool.query(
-    `INSERT INTO meeting_artifacts (meeting_id, artifact_type, name, storage_path, content_text, metadata)
-     VALUES ($1, $2, $3, $4, $5, $6)
-     RETURNING id`,
-    [params.meetingId, params.artifactType, params.name,
-     params.storagePath, params.contentText, JSON.stringify(params.metadata || {})]
-  );
-  return result.rows[0].id;
-}
-
-// ── Insights ────────────────────────────────────────────────────────────────
-
-export async function saveInsight(params: {
-  meetingId: string;
-  insightType: 'summary' | 'action_items' | 'key_topics' | 'sentiment' | 'coaching_notes';
-  content: string;
-  generatedBy?: string;
-}): Promise<string> {
-  const result = await pool.query(
-    `INSERT INTO meeting_insights (meeting_id, insight_type, content, generated_by)
-     VALUES ($1, $2, $3, $4)
-     RETURNING id`,
-    [params.meetingId, params.insightType, params.content, params.generatedBy || 'system']
-  );
-  return result.rows[0].id;
-}
-
-// ── Meeting History ──────────────────────────────────────────────────────────
-
-export async function releaseMeeting(meetingId: string): Promise<void> {
-  await pool.query(
-    'UPDATE meetings SET released_at = NOW() WHERE id = $1',
-    [meetingId]
-  );
-}
-
-export async function getMeetingsForClient(
-  clientEmail: string,
-  onlyReleased = false
-): Promise<Meeting[]> {
-  await initMeetingProjectLink();
-  const baseSelect = `
-    SELECT m.id, m.customer_id as "customerId", m.status, m.released_at,
-           m.meeting_type as "meetingType",
-           m.scheduled_at as "scheduledAt",
-           m.created_at   as "createdAt",
-           m.project_id as "projectId", p.title as "projectName"
-    FROM meetings m
-    JOIN customers c ON m.customer_id = c.id
-    LEFT JOIN tickets.tickets p ON m.project_id = p.id
-    WHERE c.email = $1`;
-
-  const query = onlyReleased
-    ? `${baseSelect} AND m.released_at IS NOT NULL ORDER BY m.created_at DESC`
-    : `${baseSelect} ORDER BY m.created_at DESC`;
-
-  const result = await pool.query(query, [clientEmail]);
-  return result.rows;
-}
-
-// ── Admin: Meeting list ──────────────────────────────────────────────────────
-
-export interface AdminMeeting {
-  id: string;
-  meetingType: string;
-  status: string;
-  talkRoomToken: string | null;
-  startedAt: Date | null;
-  endedAt: Date | null;
-  createdAt: Date;
-  customerName: string;
-  customerEmail: string;
-  customerId: string;
-  projectName: string | null;
-  projectId: string | null;
-  hasTranscript: boolean;
-  artifactCount: number;
-}
-
-export async function listAllMeetings(opts?: {
-  unassignedOnly?: boolean;
-  limit?: number;
-}): Promise<AdminMeeting[]> {
-  await initMeetingProjectLink();
-  const where = opts?.unassignedOnly
-    ? `WHERE c.name LIKE '%@unknown.local%' OR m.meeting_type = 'Talk-Session'`
-    : '';
-  const result = await pool.query(`
-    SELECT m.id, m.meeting_type AS "meetingType", m.status,
-           m.talk_room_token AS "talkRoomToken",
-           m.started_at AS "startedAt", m.ended_at AS "endedAt",
-           m.created_at AS "createdAt",
-           c.name AS "customerName", c.email AS "customerEmail",
-           c.id AS "customerId",
-           p.title AS "projectName", p.id AS "projectId",
-           EXISTS(SELECT 1 FROM transcripts t WHERE t.meeting_id = m.id) AS "hasTranscript",
-           (SELECT COUNT(*) FROM meeting_artifacts a WHERE a.meeting_id = m.id)::int AS "artifactCount"
-    FROM meetings m
-    JOIN customers c ON m.customer_id = c.id
-    LEFT JOIN tickets.tickets p ON m.project_id = p.id
-    ${where}
-    ORDER BY m.created_at DESC
-    LIMIT $1
-  `, [opts?.limit ?? 200]);
-  return result.rows;
-}
-
-export async function getMeetingDetail(meetingId: string): Promise<{
-  id: string;
-  meetingType: string;
-  status: string;
-  talkRoomToken: string | null;
-  startedAt: Date | null;
-  endedAt: Date | null;
-  createdAt: Date;
-  customerName: string;
-  customerEmail: string;
-  customerId: string;
-  projectName: string | null;
-  projectId: string | null;
-  transcript: { id: string; fullText: string } | null;
-  artifacts: Array<{ id: string; artifactType: string; name: string; storagePath: string | null; contentText: string | null }>;
-} | null> {
-  await initMeetingProjectLink();
-  const r = await pool.query(`
-    SELECT m.id, m.meeting_type AS "meetingType", m.status,
-           m.talk_room_token AS "talkRoomToken",
-           m.started_at AS "startedAt", m.ended_at AS "endedAt",
-           m.created_at AS "createdAt",
-           c.name AS "customerName", c.email AS "customerEmail", c.id AS "customerId",
-           p.title AS "projectName", p.id AS "projectId"
-    FROM meetings m
-    JOIN customers c ON m.customer_id = c.id
-    LEFT JOIN tickets.tickets p ON m.project_id = p.id
-    WHERE m.id = $1
-  `, [meetingId]);
-  if (!r.rows[0]) return null;
-  const m = r.rows[0];
-
-  const [tRow, aRows] = await Promise.all([
-    pool.query(`SELECT id, full_text AS "fullText" FROM transcripts WHERE meeting_id = $1 LIMIT 1`, [meetingId]),
-    pool.query(`SELECT id, artifact_type AS "artifactType", name, storage_path AS "storagePath", content_text AS "contentText" FROM meeting_artifacts WHERE meeting_id = $1 ORDER BY created_at`, [meetingId]),
-  ]);
-
-  return {
-    ...m,
-    transcript: tRow.rows[0] ?? null,
-    artifacts: aRows.rows,
-  };
-}
+// Meetings domain extracted to meetings-db.ts (G-SIZE03)
+// Temporary re-exports — will be removed after all callers update imports.
+export {
+  initMeetingsDb,
+  getMeetingByRoomToken,
+  createMeeting,
+  updateMeetingStatus,
+  saveTranscript,
+  saveArtifact,
+  saveInsight,
+  releaseMeeting,
+  getMeetingsForClient,
+  listAllMeetings,
+  getMeetingDetail,
+} from './meetings-db';
+export type {
+  Meeting,
+  MeetingWithDetails,
+  MeetingWithCustomer,
+  SavedTranscript,
+  AdminMeeting,
+} from './meetings-db';
 
 export async function assignMeeting(meetingId: string, params: {
   customerName?: string;
@@ -2462,125 +2136,19 @@ export async function deleteFollowUp(id: string): Promise<void> {
   await pool.query('DELETE FROM follow_ups WHERE id = $1', [id]);
 }
 
-// ── Task Calendar ─────────────────────────────────────────────────────────────
-
-export interface CalendarTask {
-  id: string;
-  name: string;
-  projectId: string;
-  projectName: string;
-  dueDate: Date;
-  status: string;
-  priority: string;
-}
-
-export async function listTasksInMonth(year: number, month: number): Promise<CalendarTask[]> {
-  await initTicketsSchema();
-  const firstDay = `${year}-${String(month).padStart(2, '0')}-01`;
-  const lastDay = new Date(Date.UTC(year, month, 0)).toISOString().slice(0, 10);
-  const result = await pool.query(
-    `SELECT pt.id,
-            pt.title AS name,
-            COALESCE(parent.parent_id, pt.parent_id) AS "projectId",
-            COALESCE(root.title, parent.title)       AS "projectName",
-            pt.due_date AS "dueDate",
-            (${STATUS_BACK_SQL.replace(/__TBL__/g, 'pt')}) AS status,
-            pt.priority
-     FROM tickets.tickets pt
-     LEFT JOIN tickets.tickets parent ON parent.id = pt.parent_id
-     LEFT JOIN tickets.tickets root   ON root.id   = parent.parent_id
-     WHERE pt.type='task'
-       AND pt.due_date BETWEEN $1::date AND $2::date
-     ORDER BY pt.due_date ASC, pt.priority DESC`,
-    [firstDay, lastDay]
-  );
-  return result.rows;
-}
-
-export interface CalendarProject {
-  id: string;
-  name: string;
-  status: string;
-  priority: string;
-  customerId: string | null;
-  customerName: string | null;
-  startDate: Date | null;
-  dueDate: Date | null;
-}
-
-export async function listProjectsInMonth(year: number, month: number, brand?: string): Promise<CalendarProject[]> {
-  await initTicketsSchema();
-  const firstDay = `${year}-${String(month).padStart(2, '0')}-01`;
-  const lastDay = new Date(Date.UTC(year, month, 0)).toISOString().slice(0, 10);
-  const result = await pool.query<CalendarProject>(
-    `SELECT p.id,
-            p.title AS name,
-            (${STATUS_BACK_SQL.replace(/__TBL__/g, 'p')}) AS status,
-            p.priority,
-            p.customer_id AS "customerId",
-            c.name        AS "customerName",
-            p.start_date  AS "startDate",
-            p.due_date    AS "dueDate"
-     FROM tickets.tickets p
-     LEFT JOIN customers c ON c.id = p.customer_id
-     WHERE p.type='project' AND p.parent_id IS NULL
-       AND p.status NOT IN ('archived', 'done')
-       AND ($1::text IS NULL OR p.brand = $1)
-       AND (
-         (p.start_date BETWEEN $2::date AND $3::date)
-         OR (p.due_date BETWEEN $2::date AND $3::date)
-       )
-     ORDER BY COALESCE(p.start_date, p.due_date) ASC`,
-    [brand ?? null, firstDay, lastDay]
-  );
-  return result.rows;
-}
-
-// ── Calendar Meetings (range query) ───────────────────────────────────────────
-
-export interface CalendarMeeting {
-  id: string;
-  meetingType: string;
-  status: string;
-  scheduledAt: Date;
-  customerId: string;
-  customerName: string;
-  customerEmail: string;
-  talkRoomToken: string | null;
-  projectId: string | null;
-  projectName: string | null;
-}
-
-/**
- * Returns all meetings whose `scheduled_at` falls inside the inclusive range
- * `[fromIso, toIso]`. Used by the admin calendar to render meetings alongside
- * tasks, projects and CalDAV bookings (T000161, T000164, T000167).
- */
-export async function listMeetingsInRange(fromIso: string, toIso: string): Promise<CalendarMeeting[]> {
-  await initMeetingProjectLink();
-  const result = await pool.query<CalendarMeeting>(
-    `SELECT m.id,
-            m.meeting_type AS "meetingType",
-            m.status,
-            m.scheduled_at AS "scheduledAt",
-            m.customer_id  AS "customerId",
-            m.talk_room_token AS "talkRoomToken",
-            c.name  AS "customerName",
-            c.email AS "customerEmail",
-            p.id    AS "projectId",
-            p.title AS "projectName"
-       FROM meetings m
-       JOIN customers c ON m.customer_id = c.id
-       LEFT JOIN tickets.tickets p ON m.project_id = p.id
-      WHERE m.scheduled_at IS NOT NULL
-        AND m.scheduled_at >= $1::timestamptz
-        AND m.scheduled_at <= $2::timestamptz
-        AND m.status NOT IN ('cancelled')
-      ORDER BY m.scheduled_at ASC`,
-    [fromIso, toIso]
-  );
-  return result.rows;
-}
+// Appointments/Calendar domain extracted to appointments-db.ts (G-SIZE03)
+// Temporary re-export so callers that still import from website-db keep working
+// until Task 2 updates their imports — removed after all callers are updated.
+export {
+  listTasksInMonth,
+  listProjectsInMonth,
+  listMeetingsInRange,
+} from './appointments-db';
+export type {
+  CalendarTask,
+  CalendarProject,
+  CalendarMeeting,
+} from './appointments-db';
 
 // ── Bug Ticket List ───────────────────────────────────────────────────────────
 
@@ -2600,311 +2168,30 @@ export interface BugTicketRow {
   fixedAt?: Date | null;
 }
 
-let bookingProjectLinksReady = false;
-async function initBookingProjectLinks(): Promise<void> {
-  if (bookingProjectLinksReady) return;
-  await pool.query(`
-    CREATE TABLE IF NOT EXISTS booking_project_links (
-      caldav_uid  TEXT    NOT NULL,
-      brand       TEXT REFERENCES public.brands(id) ON UPDATE CASCADE ON DELETE RESTRICT    NOT NULL,
-      project_id  UUID    REFERENCES tickets.tickets(id) ON DELETE SET NULL,
-      created_at  TIMESTAMPTZ NOT NULL DEFAULT now(),
-      PRIMARY KEY (caldav_uid, brand)
-    );
-      DO $$
-      BEGIN
-        IF NOT EXISTS (SELECT 1 FROM pg_constraint WHERE conname = 'booking_project_links_brand_fkey') THEN
-          ALTER TABLE booking_project_links ADD CONSTRAINT booking_project_links_brand_fkey FOREIGN KEY (brand) REFERENCES public.brands(id) ON UPDATE CASCADE ON DELETE RESTRICT NOT VALID;
-        END IF;
-      END $$;
-  `);
-  await pool.query(`
-    ALTER TABLE booking_project_links ADD COLUMN IF NOT EXISTS leistung_key TEXT
-  `);
-  bookingProjectLinksReady = true;
-}
-
-// ── Booking-Invoice Mapping ───────────────────────────────────────────────────
-
-let bookingInvoiceLinksReady = false;
-async function initBookingInvoiceLinksTable(): Promise<void> {
-  if (bookingInvoiceLinksReady) return;
-  await pool.query(`
-    CREATE TABLE IF NOT EXISTS booking_invoice_links (
-      caldav_uid      TEXT        NOT NULL,
-      brand           TEXT REFERENCES public.brands(id) ON UPDATE CASCADE ON DELETE RESTRICT        NOT NULL,
-      invoice_id      TEXT        NOT NULL,
-      invoice_number  TEXT        NOT NULL,
-      amount          NUMERIC(10,2) NOT NULL,
-      created_at      TIMESTAMPTZ NOT NULL DEFAULT now(),
-      PRIMARY KEY (caldav_uid, brand)
-    );
-      DO $$
-      BEGIN
-        IF NOT EXISTS (SELECT 1 FROM pg_constraint WHERE conname = 'booking_invoice_links_brand_fkey') THEN
-          ALTER TABLE booking_invoice_links ADD CONSTRAINT booking_invoice_links_brand_fkey FOREIGN KEY (brand) REFERENCES public.brands(id) ON UPDATE CASCADE ON DELETE RESTRICT NOT VALID;
-        END IF;
-      END $$;
-  `);
-  bookingInvoiceLinksReady = true;
-}
-
-export async function setBookingInvoice(
-  caldavUid: string,
-  brand: string,
-  invoiceId: string,
-  invoiceNumber: string,
-  amount: number
-): Promise<void> {
-  await initBookingInvoiceLinksTable();
-  await pool.query(
-    `INSERT INTO booking_invoice_links (caldav_uid, brand, invoice_id, invoice_number, amount)
-     VALUES ($1, $2, $3, $4, $5)
-     ON CONFLICT (caldav_uid, brand) DO UPDATE
-       SET invoice_id = EXCLUDED.invoice_id,
-           invoice_number = EXCLUDED.invoice_number,
-           amount = EXCLUDED.amount`,
-    [caldavUid, brand, invoiceId, invoiceNumber, amount]
-  );
-}
-
-export interface BookingInvoiceInfo {
-  invoiceId: string;
-  invoiceNumber: string;
-  amount: number;
-}
-
-export async function getBookingInvoices(caldavUids: string[], brand: string): Promise<Map<string, BookingInvoiceInfo>> {
-  if (caldavUids.length === 0) return new Map();
-  await initBookingInvoiceLinksTable();
-  const result = await pool.query(
-    `SELECT caldav_uid, invoice_id, invoice_number, amount
-     FROM booking_invoice_links
-     WHERE caldav_uid = ANY($1) AND brand = $2`,
-    [caldavUids, brand]
-  );
-  return new Map(
-    result.rows.map((r: {
-      caldav_uid: string;
-      invoice_id: string;
-      invoice_number: string;
-      amount: string;
-    }) => [
-      r.caldav_uid,
-      {
-        invoiceId: r.invoice_id,
-        invoiceNumber: r.invoice_number,
-        amount: parseFloat(r.amount),
-      },
-    ])
-  );
-}
-
-export async function getBookingProjects(caldavUids: string[], brand: string): Promise<Map<string, string>> {
-  if (caldavUids.length === 0) return new Map();
-  await initBookingProjectLinks();
-  const result = await pool.query(
-    `SELECT caldav_uid, project_id FROM booking_project_links
-     WHERE caldav_uid = ANY($1) AND brand = $2 AND project_id IS NOT NULL`,
-    [caldavUids, brand]
-  );
-  return new Map(result.rows.map((r: { caldav_uid: string; project_id: string }) => [r.caldav_uid, r.project_id]));
-}
-
-export async function setBookingProject(
-  caldavUid: string,
-  projectId: string | null,
-  brand: string,
-  leistungKey?: string
-): Promise<void> {
-  await initBookingProjectLinks();
-  if (!projectId) {
-    await pool.query(
-      `DELETE FROM booking_project_links WHERE caldav_uid = $1 AND brand = $2`,
-      [caldavUid, brand]
-    );
-  } else {
-    await pool.query(
-      `INSERT INTO booking_project_links (caldav_uid, brand, project_id, leistung_key)
-       VALUES ($1, $2, $3, $4)
-       ON CONFLICT (caldav_uid, brand) DO UPDATE
-         SET project_id  = EXCLUDED.project_id,
-             leistung_key = EXCLUDED.leistung_key`,
-      [caldavUid, brand, projectId, leistungKey ?? null]
-    );
-  }
-}
-
-export async function getBookingLeistungen(caldavUids: string[], brand: string): Promise<Map<string, string>> {
-  if (caldavUids.length === 0) return new Map();
-  await initBookingProjectLinks();
-  const result = await pool.query(
-    `SELECT caldav_uid, leistung_key FROM booking_project_links
-     WHERE caldav_uid = ANY($1) AND brand = $2 AND leistung_key IS NOT NULL`,
-    [caldavUids, brand]
-  );
-  return new Map(result.rows.map((r: { caldav_uid: string; leistung_key: string }) => [r.caldav_uid, r.leistung_key]));
-}
-
-// ── Slot Whitelist ────────────────────────────────────────────────────────────
-
-export interface WhitelistedSlot {
-  slotStart: Date;
-  slotEnd: Date;
-}
-
-async function initSlotWhitelistTable(): Promise<void> {
-  await pool.query(`
-    CREATE TABLE IF NOT EXISTS slot_whitelist (
-      brand      TEXT REFERENCES public.brands(id) ON UPDATE CASCADE ON DELETE RESTRICT        NOT NULL,
-      slot_start TIMESTAMPTZ NOT NULL,
-      slot_end   TIMESTAMPTZ NOT NULL,
-      created_at TIMESTAMPTZ NOT NULL DEFAULT now(),
-      PRIMARY KEY (brand, slot_start)
-    );
-      DO $$
-      BEGIN
-        IF NOT EXISTS (SELECT 1 FROM pg_constraint WHERE conname = 'slot_whitelist_brand_fkey') THEN
-          ALTER TABLE slot_whitelist ADD CONSTRAINT slot_whitelist_brand_fkey FOREIGN KEY (brand) REFERENCES public.brands(id) ON UPDATE CASCADE ON DELETE RESTRICT NOT VALID;
-        END IF;
-      END $$;
-  `);
-}
-
-export async function getWhitelistedSlots(brand: string): Promise<WhitelistedSlot[]> {
-  await initSlotWhitelistTable();
-  // Only return future slots for display — isSlotWhitelisted has no time filter
-  // so booking validation works even for slots that just started.
-  const result = await pool.query(
-    `SELECT slot_start AS "slotStart", slot_end AS "slotEnd"
-     FROM slot_whitelist
-     WHERE brand = $1 AND slot_start > now()
-     ORDER BY slot_start ASC`,
-    [brand]
-  );
-  return result.rows;
-}
-
-export async function addSlotToWhitelist(brand: string, start: Date, end: Date): Promise<void> {
-  await initSlotWhitelistTable();
-  await pool.query(
-    `INSERT INTO slot_whitelist (brand, slot_start, slot_end)
-     VALUES ($1, $2, $3)
-     ON CONFLICT (brand, slot_start) DO UPDATE SET slot_end = $3`,
-    [brand, start, end]
-  );
-}
-
-export async function removeSlotFromWhitelist(brand: string, start: Date): Promise<void> {
-  await initSlotWhitelistTable();
-  await pool.query(
-    'DELETE FROM slot_whitelist WHERE brand = $1 AND slot_start = $2::timestamptz',
-    [brand, start]
-  );
-}
-
-export async function isSlotWhitelisted(brand: string, start: Date): Promise<boolean> {
-  await initSlotWhitelistTable();
-  const result = await pool.query(
-    'SELECT 1 FROM slot_whitelist WHERE brand = $1 AND slot_start = $2::timestamptz',
-    [brand, start]
-  );
-  return (result.rowCount ?? 0) > 0;
-}
-
-// Atomically removes the slot from the whitelist and returns true if it was
-// available (i.e. not already claimed by another concurrent booking).
-export async function claimSlot(brand: string, start: Date): Promise<boolean> {
-  await initSlotWhitelistTable();
-  const result = await pool.query(
-    'DELETE FROM slot_whitelist WHERE brand = $1 AND slot_start = $2::timestamptz RETURNING 1',
-    [brand, start]
-  );
-  return (result.rowCount ?? 0) > 0;
-}
-
-// ── Free Time Windows ────────────────────────────────────────────────────────
-
-export interface FreeTimeWindow {
-  id: string;
-  date: string;     // YYYY-MM-DD
-  winStart: string; // HH:MM
-  winEnd: string;   // HH:MM
-}
-
-async function initFreeTimeWindowsTable(): Promise<void> {
-  await pool.query(`
-    CREATE TABLE IF NOT EXISTS free_time_windows (
-      id         TEXT        NOT NULL DEFAULT gen_random_uuid()::text,
-      brand      TEXT REFERENCES public.brands(id) ON UPDATE CASCADE ON DELETE RESTRICT        NOT NULL,
-      date       DATE        NOT NULL,
-      win_start  TIME        NOT NULL,
-      win_end    TIME        NOT NULL,
-      created_at TIMESTAMPTZ NOT NULL DEFAULT now(),
-      PRIMARY KEY (id)
-    );
-      DO $$
-      BEGIN
-        IF NOT EXISTS (SELECT 1 FROM pg_constraint WHERE conname = 'free_time_windows_brand_fkey') THEN
-          ALTER TABLE free_time_windows ADD CONSTRAINT free_time_windows_brand_fkey FOREIGN KEY (brand) REFERENCES public.brands(id) ON UPDATE CASCADE ON DELETE RESTRICT NOT VALID;
-        END IF;
-      END $$;
-  `);
-}
-
-export async function getFreeTimeWindows(brand: string, fromDate?: string, toDate?: string): Promise<FreeTimeWindow[]> {
-  await initFreeTimeWindowsTable();
-  const result = await pool.query(
-    `SELECT id,
-            to_char(date, 'YYYY-MM-DD')   AS date,
-            to_char(win_start, 'HH24:MI') AS "winStart",
-            to_char(win_end,   'HH24:MI') AS "winEnd"
-     FROM free_time_windows
-     WHERE brand = $1
-       AND ($2::date IS NULL OR date >= $2::date)
-       AND ($3::date IS NULL OR date <= $3::date)
-     ORDER BY date ASC, win_start ASC`,
-    [brand, fromDate ?? null, toDate ?? null]
-  );
-  return result.rows;
-}
-
-export async function addFreeTimeWindow(brand: string, date: string, winStart: string, winEnd: string): Promise<string> {
-  await initFreeTimeWindowsTable();
-  const result = await pool.query(
-    `INSERT INTO free_time_windows (brand, date, win_start, win_end)
-     VALUES ($1, $2::date, $3::time, $4::time)
-     RETURNING id`,
-    [brand, date, winStart, winEnd]
-  );
-  return result.rows[0].id as string;
-}
-
-export async function removeFreeTimeWindow(brand: string, id: string): Promise<void> {
-  await initFreeTimeWindowsTable();
-  await pool.query(
-    'DELETE FROM free_time_windows WHERE id = $1 AND brand = $2',
-    [id, brand]
-  );
-}
-
-export async function isSlotInAnyWindow(brand: string, slotStart: Date, slotEnd: Date): Promise<boolean> {
-  await initFreeTimeWindowsTable();
-  const dateStr = slotStart.toISOString().split('T')[0];
-  const sh = slotStart.getHours().toString().padStart(2, '0');
-  const sm = slotStart.getMinutes().toString().padStart(2, '0');
-  const eh = slotEnd.getHours().toString().padStart(2, '0');
-  const em = slotEnd.getMinutes().toString().padStart(2, '0');
-  const result = await pool.query(
-    `SELECT 1 FROM free_time_windows
-     WHERE brand = $1
-       AND date = $2::date
-       AND win_start <= $3::time
-       AND win_end   >= $4::time`,
-    [brand, dateStr, `${sh}:${sm}`, `${eh}:${em}`]
-  );
-  return (result.rowCount ?? 0) > 0;
-}
+// Booking/slot/time-window domain extracted to appointments-db.ts (G-SIZE03)
+// Temporary re-exports so callers that still import from website-db keep working
+// until Task 2 updates their imports — removed after all callers are updated.
+export {
+  setBookingInvoice,
+  getBookingInvoices,
+  getBookingProjects,
+  setBookingProject,
+  getBookingLeistungen,
+  getWhitelistedSlots,
+  addSlotToWhitelist,
+  removeSlotFromWhitelist,
+  isSlotWhitelisted,
+  claimSlot,
+  getFreeTimeWindows,
+  addFreeTimeWindow,
+  removeFreeTimeWindow,
+  isSlotInAnyWindow,
+} from './appointments-db';
+export type {
+  BookingInvoiceInfo,
+  WhitelistedSlot,
+  FreeTimeWindow,
+} from './appointments-db';
 
 export async function listBugTickets(filters: {
   status?: string;
@@ -3280,248 +2567,16 @@ export async function claimBrettLinkPost(meetingId: string): Promise<boolean> {
   return result.rowCount === 1;
 }
 
-// ── Test Runs ────────────────────────────────────────────────────────────────
-
-export interface TestRun {
-  id: string;
-  tier: string;
-  testIds: string | null;
-  cluster: string;
-  startedAt: string;
-  finishedAt: string | null;
-  status: 'running' | 'done' | 'error';
-  pass: number | null;
-  fail: number | null;
-  skip: number | null;
-  durationMs: number | null;
-}
-
-async function initTestRunsTable(): Promise<void> {
-  await pool.query(`
-    CREATE TABLE IF NOT EXISTS test_runs (
-      id           TEXT PRIMARY KEY,
-      tier         TEXT NOT NULL,
-      test_ids     TEXT,
-      cluster      TEXT NOT NULL,
-      started_at   TIMESTAMPTZ NOT NULL DEFAULT now(),
-      finished_at  TIMESTAMPTZ,
-      status       TEXT NOT NULL DEFAULT 'running',
-      pass         INT,
-      fail         INT,
-      skip         INT,
-      duration_ms  INT
-    )
-  `);
-}
-
-export async function saveTestRun(params: {
-  id: string;
-  tier: string;
-  testIds: string | null;
-  cluster: string;
-}): Promise<void> {
-  await initTestRunsTable();
-  await pool.query(
-    `INSERT INTO test_runs (id, tier, test_ids, cluster) VALUES ($1, $2, $3, $4)`,
-    [params.id, params.tier, params.testIds, params.cluster]
-  );
-}
-
-export async function updateTestRun(params: {
-  id: string;
-  status: 'done' | 'error';
-  pass: number;
-  fail: number;
-  skip: number;
-  durationMs: number;
-}): Promise<void> {
-  await pool.query(
-    `UPDATE test_runs
-     SET status = $2, finished_at = now(), pass = $3, fail = $4, skip = $5, duration_ms = $6
-     WHERE id = $1`,
-    [params.id, params.status, params.pass, params.fail, params.skip, params.durationMs]
-  );
-}
-
-export async function listTestRuns(limit = 20): Promise<TestRun[]> {
-  await initTestRunsTable();
-  const result = await pool.query(
-    `SELECT id, tier, test_ids AS "testIds", cluster,
-            started_at AS "startedAt", finished_at AS "finishedAt",
-            status, pass, fail, skip, duration_ms AS "durationMs"
-     FROM test_runs ORDER BY started_at DESC LIMIT $1`,
-    [limit]
-  );
-  return result.rows;
-}
-
-// ── Test Results (per-test history for flake detection + trends) ─────────────
-
-export interface TestResultRow {
-  testId: string;
-  category: 'FA' | 'SA' | 'NFA' | 'AK' | 'E2E' | 'BATS';
-  status: 'pass' | 'fail' | 'skip';
-  durationMs?: number;
-  message?: string;
-}
-
-export interface SavedTestResult {
-  id: number;
-  testId: string;
-  category: string;
-  status: string;
-  durationMs: number | null;
-  message: string | null;
-}
-
-export async function saveTestResults(
-  runId: string,
-  rows: TestResultRow[],
-): Promise<SavedTestResult[]> {
-  if (rows.length === 0) return [];
-  const values: unknown[] = [];
-  const placeholders = rows.map((r, i) => {
-    const base = i * 6;
-    values.push(runId, r.testId, r.category, r.status, r.durationMs ?? null, r.message ?? null);
-    return `($${base + 1}, $${base + 2}, $${base + 3}, $${base + 4}, $${base + 5}, $${base + 6})`;
-  }).join(',');
-  const result = await pool.query<{
-    id: number; test_id: string; category: string; status: string;
-    duration_ms: number | null; message: string | null;
-  }>(
-    `INSERT INTO test_results (run_id, test_id, category, status, duration_ms, message)
-     VALUES ${placeholders}
-     RETURNING id, test_id, category, status, duration_ms, message`,
-    values,
-  );
-  return result.rows.map(r => ({
-    id: r.id,
-    testId: r.test_id,
-    category: r.category,
-    status: r.status,
-    durationMs: r.duration_ms,
-    message: r.message,
-  }));
-}
-
-export interface FlakeRow {
-  testId: string;
-  category: string;
-  recentRuns: Array<{ runId: string; status: string; createdAt: string }>;
-  failureRate: number;
-}
-
-export async function listFlakeWindow(limit: number): Promise<FlakeRow[]> {
-  const result = await pool.query<{
-    test_id: string;
-    category: string;
-    recent: Array<{ run_id: string; status: string; created_at: string }>;
-  }>(
-    `WITH ranked AS (
-       SELECT test_id, category, run_id, status, created_at,
-              row_number() OVER (PARTITION BY test_id ORDER BY created_at DESC) AS rn
-         FROM test_results
-     )
-     SELECT test_id, category,
-            jsonb_agg(jsonb_build_object('run_id', run_id, 'status', status, 'created_at', created_at) ORDER BY created_at DESC) AS recent
-       FROM ranked
-      WHERE rn <= $1
-   GROUP BY test_id, category`,
-    [limit],
-  );
-  return result.rows.map(row => {
-    const recent = row.recent.map(r => ({ runId: r.run_id, status: r.status, createdAt: r.created_at }));
-    const fails = recent.filter(r => r.status === 'fail').length;
-    return {
-      testId: row.test_id,
-      category: row.category,
-      recentRuns: recent,
-      failureRate: recent.length === 0 ? 0 : fails / recent.length,
-    };
-  }).sort((a, b) => b.failureRate - a.failureRate);
-}
-
-export interface TrendRow { date: string; pass: number; fail: number; skip: number; p50DurationMs: number; p95DurationMs: number; }
-
-export async function getTestRunTrend(days: number): Promise<TrendRow[]> {
-  const result = await pool.query<{
-    day: string; pass: string; fail: string; skip: string; p50: string; p95: string;
-  }>(
-    `SELECT to_char(date_trunc('day', started_at), 'YYYY-MM-DD') AS day,
-            sum(coalesce(pass, 0))::text AS pass,
-            sum(coalesce(fail, 0))::text AS fail,
-            sum(coalesce(skip, 0))::text AS skip,
-            coalesce(percentile_cont(0.5) WITHIN GROUP (ORDER BY duration_ms), 0)::text AS p50,
-            coalesce(percentile_cont(0.95) WITHIN GROUP (ORDER BY duration_ms), 0)::text AS p95
-       FROM test_runs
-      WHERE started_at >= now() - ($1 || ' days')::interval
-   GROUP BY day
-   ORDER BY day`,
-    [days],
-  );
-  return result.rows.map(r => ({
-    date: r.day,
-    pass: Number(r.pass),
-    fail: Number(r.fail),
-    skip: Number(r.skip),
-    p50DurationMs: Number(r.p50),
-    p95DurationMs: Number(r.p95),
-  }));
-}
-
-export async function listLastTestStatusPerTest(): Promise<Array<{ testId: string; status: string; createdAt: string }>> {
-  const result = await pool.query<{ test_id: string; status: string; created_at: string }>(
-    `SELECT DISTINCT ON (test_id) test_id, status, created_at FROM test_results ORDER BY test_id, created_at DESC`,
-  );
-  return result.rows.map(r => ({ testId: r.test_id, status: r.status, createdAt: r.created_at }));
-}
-
-// ── Playwright Reports ───────────────────────────────────────────────────────
-
-export interface PlaywrightReport {
-  id: number;
-  createdAt: string;
-  html: string;
-}
-
-async function initPlaywrightReportsTable(): Promise<void> {
-  await pool.query(`
-    CREATE TABLE IF NOT EXISTS playwright_reports (
-      id         SERIAL PRIMARY KEY,
-      created_at TIMESTAMPTZ NOT NULL DEFAULT now(),
-      html       TEXT NOT NULL
-    )
-  `);
-}
-
-export async function savePlaywrightReport(html: string): Promise<number> {
-  await initPlaywrightReportsTable();
-  const result = await pool.query(
-    `INSERT INTO playwright_reports (html) VALUES ($1) RETURNING id`,
-    [html]
-  );
-  // Keep only last 5
-  await pool.query(
-    `DELETE FROM playwright_reports WHERE id NOT IN (
-       SELECT id FROM playwright_reports ORDER BY created_at DESC LIMIT 5
-     )`
-  );
-  return result.rows[0].id;
-}
-
-export async function getLatestPlaywrightReport(): Promise<PlaywrightReport | null> {
-  await initPlaywrightReportsTable();
-  const result = await pool.query(
-    `SELECT id, created_at AS "createdAt", html
-     FROM playwright_reports ORDER BY created_at DESC LIMIT 1`
-  );
-  if (result.rows.length === 0) return null;
-  return {
-    id: result.rows[0].id,
-    createdAt: result.rows[0].createdAt.toISOString(),
-    html: result.rows[0].html,
-  };
-}
+// Test infra domain extracted to test-infra-db.ts (G-SIZE03)
+// Temporary re-exports — will be removed after callers update imports.
+export {
+  saveTestRun, updateTestRun, listTestRuns,
+  saveTestResults, listFlakeWindow, getTestRunTrend, listLastTestStatusPerTest,
+  savePlaywrightReport, getLatestPlaywrightReport,
+} from './test-infra-db';
+export type {
+  TestRun, TestResultRow, SavedTestResult, FlakeRow, TrendRow, PlaywrightReport,
+} from './test-infra-db';
 
 // ── Custom Website Sections ────────────────────────────────────────────────
 
@@ -3627,613 +2682,12 @@ export async function deleteCustomSection(slug: string): Promise<void> {
   await pool.query('DELETE FROM website_custom_sections WHERE slug = $1', [slug]);
 }
 
-// ── Billing Tables ───────────────────────────────────────────────────────────
+// Billing schema-init domain extracted to billing-db.ts (G-SIZE03)
+// Temporary re-exports — will be removed after callers update imports.
+export { initBillingTables, initTaxMonitorTables, initEurTables } from './billing-db';
 
-async function initBillingAuditTable(): Promise<void> {
-  await pool.query(`
-    CREATE TABLE IF NOT EXISTS billing_audit_log (
-      id            BIGSERIAL PRIMARY KEY,
-      invoice_id    TEXT NOT NULL REFERENCES billing_invoices(id),
-      action        TEXT NOT NULL,
-      actor_user_id TEXT,
-      actor_email   TEXT,
-      from_status   TEXT,
-      to_status     TEXT,
-      reason        TEXT,
-      metadata      JSONB,
-      created_at    TIMESTAMPTZ NOT NULL DEFAULT now()
-    )
-  `);
-  await pool.query(`CREATE INDEX IF NOT EXISTS idx_billing_audit_invoice ON billing_audit_log(invoice_id, created_at DESC)`);
-}
+// ── (end of billing — was lines 2912-3519, extracted to billing-db.ts) ───────
 
-async function installInvoiceImmutabilityTriggers(): Promise<void> {
-  try {
-    await installInvoiceImmutabilityTriggersInner();
-  } catch (err) {
-    // 42501 = insufficient_privilege. Triggers/functions exist from a prior
-    // deploy under a different role (e.g. postgres superuser); current role
-    // can't replace them but they enforce the same invariants. Leaving them
-    // alone is correct — `initBillingTables` runs on every billing call so a
-    // hard error here would break the entire billing API in production.
-    if ((err as { code?: string } | null)?.code === '42501') return;
-    throw err;
-  }
-}
-
-async function installInvoiceImmutabilityTriggersInner(): Promise<void> {
-  await pool.query(`
-    CREATE OR REPLACE FUNCTION billing_invoices_immutable() RETURNS trigger AS $fn$
-    BEGIN
-      IF OLD.locked = true THEN
-        IF NEW.net_amount   IS DISTINCT FROM OLD.net_amount   OR
-           NEW.tax_rate     IS DISTINCT FROM OLD.tax_rate     OR
-           NEW.tax_amount   IS DISTINCT FROM OLD.tax_amount   OR
-           NEW.gross_amount IS DISTINCT FROM OLD.gross_amount OR
-           NEW.tax_mode     IS DISTINCT FROM OLD.tax_mode     OR
-           NEW.customer_id  IS DISTINCT FROM OLD.customer_id  OR
-           NEW.issue_date   IS DISTINCT FROM OLD.issue_date   OR
-           NEW.due_date     IS DISTINCT FROM OLD.due_date     OR
-           NEW.number       IS DISTINCT FROM OLD.number       OR
-           NEW.brand        IS DISTINCT FROM OLD.brand        OR
-           (OLD.hash_sha256 IS NOT NULL AND NEW.hash_sha256 IS DISTINCT FROM OLD.hash_sha256)
-        THEN
-          RAISE EXCEPTION 'GoBD: locked invoice % cannot be modified', OLD.id;
-        END IF;
-      END IF;
-      RETURN NEW;
-    END $fn$ LANGUAGE plpgsql;
-  `);
-  await pool.query(`
-    CREATE OR REPLACE FUNCTION billing_invoices_no_delete() RETURNS trigger AS $fn$
-    BEGIN
-      IF OLD.locked = true THEN
-        RAISE EXCEPTION 'GoBD: locked invoice % cannot be deleted', OLD.id;
-      END IF;
-      RETURN OLD;
-    END $fn$ LANGUAGE plpgsql;
-  `);
-  await pool.query(`
-    CREATE OR REPLACE FUNCTION billing_lines_immutable() RETURNS trigger AS $fn$
-    DECLARE inv_locked boolean;
-    BEGIN
-      SELECT locked INTO inv_locked FROM billing_invoices
-        WHERE id = COALESCE(NEW.invoice_id, OLD.invoice_id);
-      IF inv_locked = true THEN
-        RAISE EXCEPTION 'GoBD: cannot modify lines of locked invoice %', COALESCE(NEW.invoice_id, OLD.invoice_id);
-      END IF;
-      RETURN COALESCE(NEW, OLD);
-    END $fn$ LANGUAGE plpgsql;
-  `);
-  await pool.query(`DROP TRIGGER IF EXISTS billing_invoices_immutable_trg ON billing_invoices`);
-  await pool.query(`CREATE TRIGGER billing_invoices_immutable_trg
-    BEFORE UPDATE ON billing_invoices
-    FOR EACH ROW EXECUTE FUNCTION billing_invoices_immutable()`);
-  await pool.query(`DROP TRIGGER IF EXISTS billing_invoices_no_delete_trg ON billing_invoices`);
-  await pool.query(`CREATE TRIGGER billing_invoices_no_delete_trg
-    BEFORE DELETE ON billing_invoices
-    FOR EACH ROW EXECUTE FUNCTION billing_invoices_no_delete()`);
-  await pool.query(`DROP TRIGGER IF EXISTS billing_lines_immutable_trg ON billing_invoice_line_items`);
-  await pool.query(`CREATE TRIGGER billing_lines_immutable_trg
-    BEFORE INSERT OR UPDATE OR DELETE ON billing_invoice_line_items
-    FOR EACH ROW EXECUTE FUNCTION billing_lines_immutable()`);
-}
-
-let billingTablesReady = false;
-export async function initBillingTables(): Promise<void> {
-  if (billingTablesReady) return;
-  await pool.query(`
-    CREATE TABLE IF NOT EXISTS billing_customers (
-      id            TEXT PRIMARY KEY DEFAULT gen_random_uuid()::text,
-      brand         TEXT REFERENCES public.brands(id) ON UPDATE CASCADE ON DELETE RESTRICT NOT NULL,
-      name          TEXT NOT NULL,
-      email         TEXT NOT NULL,
-      company       TEXT,
-      address_line1 TEXT,
-      city          TEXT,
-      postal_code   TEXT,
-      land_iso      CHAR(2) NOT NULL DEFAULT 'DE',
-      vat_number    TEXT,
-      sepa_iban     TEXT,
-      sepa_bic      TEXT,
-      sepa_mandate_ref  TEXT,
-      sepa_mandate_date DATE,
-      created_at    TIMESTAMPTZ NOT NULL DEFAULT now(),
-      typ           TEXT NOT NULL DEFAULT 'Kunde',
-      CONSTRAINT billing_customers_brand_email_typ_key UNIQUE (brand, email, typ)
-    );
-      DO $$
-      BEGIN
-        IF NOT EXISTS (SELECT 1 FROM pg_constraint WHERE conname = 'billing_customers_brand_fkey') THEN
-          ALTER TABLE billing_customers ADD CONSTRAINT billing_customers_brand_fkey FOREIGN KEY (brand) REFERENCES public.brands(id) ON UPDATE CASCADE ON DELETE RESTRICT NOT VALID;
-        END IF;
-      END $$;
-  `);
-  await pool.query(`ALTER TABLE billing_customers ADD COLUMN IF NOT EXISTS default_leitweg_id TEXT`);
-  await pool.query(`ALTER TABLE billing_customers ADD COLUMN IF NOT EXISTS customers_id UUID REFERENCES customers(id)`);
-  await pool.query(`
-    DO $$
-    BEGIN
-      IF EXISTS (
-        SELECT 1 FROM information_schema.columns
-        WHERE table_schema='public'
-          AND table_name='billing_customers' AND column_name='country'
-      ) THEN
-        ALTER TABLE billing_customers RENAME COLUMN country TO land_iso;
-      END IF;
-    END $$
-  `);
-  await pool.query(`
-    DO $$
-    BEGIN
-      IF EXISTS (
-        SELECT 1 FROM information_schema.columns
-        WHERE table_schema='public'
-          AND table_name='billing_customers' AND column_name='land_iso' AND data_type != 'character'
-      ) THEN
-        ALTER TABLE billing_customers ALTER COLUMN land_iso TYPE CHAR(2);
-      END IF;
-    END $$
-  `);
-  await pool.query(`
-    ALTER TABLE billing_customers
-      ADD COLUMN IF NOT EXISTS typ TEXT NOT NULL DEFAULT 'Kunde'
-  `);
-  await pool.query(`
-    DO $$
-    BEGIN
-      IF NOT EXISTS (
-        SELECT 1 FROM pg_constraint WHERE conname='billing_customers_typ_chk'
-      ) THEN
-        ALTER TABLE billing_customers
-          ADD CONSTRAINT billing_customers_typ_chk CHECK (typ IN ('Kunde'));
-      END IF;
-      IF EXISTS (
-        SELECT 1 FROM pg_constraint WHERE conname='billing_customers_brand_email_key'
-      ) THEN
-        ALTER TABLE billing_customers DROP CONSTRAINT billing_customers_brand_email_key;
-      END IF;
-      IF NOT EXISTS (
-        SELECT 1 FROM pg_constraint WHERE conname='billing_customers_brand_email_typ_key'
-      ) THEN
-        ALTER TABLE billing_customers
-          ADD CONSTRAINT billing_customers_brand_email_typ_key UNIQUE (brand, email, typ);
-      END IF;
-    END $$
-  `);
-  await pool.query(`
-    CREATE TABLE IF NOT EXISTS billing_invoices (
-      id            TEXT PRIMARY KEY DEFAULT gen_random_uuid()::text,
-      brand         TEXT REFERENCES public.brands(id) ON UPDATE CASCADE ON DELETE RESTRICT NOT NULL,
-      number        TEXT NOT NULL UNIQUE,
-      status        TEXT NOT NULL DEFAULT 'draft',
-      customer_id   TEXT NOT NULL REFERENCES billing_customers(id),
-      issue_date    DATE NOT NULL,
-      due_date      DATE NOT NULL,
-      service_period_start DATE,
-      service_period_end   DATE,
-      tax_mode      TEXT NOT NULL,
-      net_amount    NUMERIC(12,2) NOT NULL,
-      tax_rate      NUMERIC(5,2)  NOT NULL DEFAULT 0,
-      tax_amount    NUMERIC(12,2) NOT NULL DEFAULT 0,
-      gross_amount  NUMERIC(12,2) NOT NULL,
-      notes         TEXT,
-      payment_reference TEXT,
-      locked        BOOLEAN NOT NULL DEFAULT false,
-      cancels_invoice_id TEXT REFERENCES billing_invoices(id),
-      retain_until  DATE NOT NULL DEFAULT (CURRENT_DATE + INTERVAL '10 years'),
-      pdf_path      TEXT,
-      created_at    TIMESTAMPTZ NOT NULL DEFAULT now(),
-      updated_at    TIMESTAMPTZ NOT NULL DEFAULT now()
-    );
-      DO $$
-      BEGIN
-        IF NOT EXISTS (SELECT 1 FROM pg_constraint WHERE conname = 'billing_invoices_brand_fkey') THEN
-          ALTER TABLE billing_invoices ADD CONSTRAINT billing_invoices_brand_fkey FOREIGN KEY (brand) REFERENCES public.brands(id) ON UPDATE CASCADE ON DELETE RESTRICT NOT VALID;
-        END IF;
-      END $$;
-  `);
-  await pool.query(`ALTER TABLE billing_invoices ADD COLUMN IF NOT EXISTS pdf_path TEXT`);
-  await pool.query(`ALTER TABLE billing_invoices ADD COLUMN IF NOT EXISTS leitweg_id TEXT`);
-  await pool.query(`ALTER TABLE billing_invoices ADD COLUMN IF NOT EXISTS einvoice_validated_at TIMESTAMPTZ`);
-  await pool.query(`ALTER TABLE billing_invoices ADD COLUMN IF NOT EXISTS einvoice_validation_report JSONB`);
-  await pool.query(`
-    ALTER TABLE billing_invoices
-      ADD COLUMN IF NOT EXISTS kind TEXT NOT NULL DEFAULT 'regular',
-      ADD COLUMN IF NOT EXISTS parent_invoice_id TEXT REFERENCES billing_invoices(id),
-      ADD COLUMN IF NOT EXISTS currency CHAR(3) NOT NULL DEFAULT 'EUR',
-      ADD COLUMN IF NOT EXISTS currency_rate NUMERIC(12,6),
-      ADD COLUMN IF NOT EXISTS net_amount_eur  NUMERIC(12,2),
-      ADD COLUMN IF NOT EXISTS gross_amount_eur NUMERIC(12,2),
-      ADD COLUMN IF NOT EXISTS supply_type     TEXT
-  `);
-  await pool.query(`
-    DO $$ BEGIN
-      IF NOT EXISTS (
-        SELECT 1 FROM pg_constraint WHERE conname='billing_invoices_kind_chk'
-      ) THEN
-        ALTER TABLE billing_invoices
-          ADD CONSTRAINT billing_invoices_kind_chk
-          CHECK (kind IN ('regular','prepayment','final','gutschrift'));
-      END IF;
-    END $$
-  `);
-  await pool.query(`
-    CREATE TABLE IF NOT EXISTS billing_invoice_dunnings (
-      id            TEXT PRIMARY KEY DEFAULT gen_random_uuid()::text,
-      invoice_id    TEXT NOT NULL REFERENCES billing_invoices(id),
-      brand         TEXT REFERENCES public.brands(id) ON UPDATE CASCADE ON DELETE RESTRICT NOT NULL,
-      level         SMALLINT NOT NULL,
-      generated_at  TIMESTAMPTZ NOT NULL DEFAULT now(),
-      sent_at       TIMESTAMPTZ,
-      sent_by       TEXT,
-      fee_amount    NUMERIC(12,2) NOT NULL DEFAULT 0,
-      interest_amount NUMERIC(12,2) NOT NULL DEFAULT 0,
-      outstanding_at_generation NUMERIC(12,2) NOT NULL,
-      pdf_path      TEXT,
-      UNIQUE (invoice_id, level)
-    );
-      DO $$
-      BEGIN
-        IF NOT EXISTS (SELECT 1 FROM pg_constraint WHERE conname = 'billing_invoice_dunnings_brand_fkey') THEN
-          ALTER TABLE billing_invoice_dunnings ADD CONSTRAINT billing_invoice_dunnings_brand_fkey FOREIGN KEY (brand) REFERENCES public.brands(id) ON UPDATE CASCADE ON DELETE RESTRICT NOT VALID;
-        END IF;
-      END $$;
-  `);
-  await pool.query(`
-    CREATE TABLE IF NOT EXISTS billing_invoice_line_items (
-      id          BIGSERIAL PRIMARY KEY,
-      invoice_id  TEXT NOT NULL REFERENCES billing_invoices(id) ON DELETE CASCADE,
-      description TEXT NOT NULL,
-      quantity    NUMERIC(10,2) NOT NULL DEFAULT 1,
-      unit        TEXT,
-      unit_price  NUMERIC(12,2) NOT NULL,
-      net_amount  NUMERIC(12,2) NOT NULL
-    )
-  `);
-  await pool.query(`
-    CREATE TABLE IF NOT EXISTS billing_invoice_payments (
-      id           BIGSERIAL PRIMARY KEY,
-      invoice_id   TEXT NOT NULL REFERENCES billing_invoices(id),
-      brand        TEXT REFERENCES public.brands(id) ON UPDATE CASCADE ON DELETE RESTRICT NOT NULL,
-      paid_at      DATE NOT NULL,
-      amount       NUMERIC(12,2) NOT NULL CHECK (amount <> 0),
-      method       TEXT NOT NULL,
-      reference    TEXT,
-      recorded_by  TEXT NOT NULL,
-      notes        TEXT,
-      created_at   TIMESTAMPTZ NOT NULL DEFAULT now()
-    );
-      DO $$
-      BEGIN
-        IF NOT EXISTS (SELECT 1 FROM pg_constraint WHERE conname = 'billing_invoice_payments_brand_fkey') THEN
-          ALTER TABLE billing_invoice_payments ADD CONSTRAINT billing_invoice_payments_brand_fkey FOREIGN KEY (brand) REFERENCES public.brands(id) ON UPDATE CASCADE ON DELETE RESTRICT NOT VALID;
-        END IF;
-      END $$;
-  `);
-  await pool.query(`
-    CREATE INDEX IF NOT EXISTS billing_invoice_payments_invoice_idx
-      ON billing_invoice_payments (invoice_id)
-  `);
-  await pool.query(`
-    CREATE TABLE IF NOT EXISTS billing_quotes (
-      id            TEXT PRIMARY KEY DEFAULT gen_random_uuid()::text,
-      brand         TEXT REFERENCES public.brands(id) ON UPDATE CASCADE ON DELETE RESTRICT NOT NULL,
-      number        TEXT NOT NULL UNIQUE,
-      status        TEXT NOT NULL DEFAULT 'draft',
-      customer_id   TEXT NOT NULL REFERENCES billing_customers(id),
-      issue_date    DATE NOT NULL,
-      valid_until   DATE,
-      net_amount    NUMERIC(12,2) NOT NULL,
-      tax_rate      NUMERIC(5,2)  NOT NULL DEFAULT 0,
-      gross_amount  NUMERIC(12,2) NOT NULL,
-      notes         TEXT,
-      converted_to_invoice_id TEXT REFERENCES billing_invoices(id),
-      created_at    TIMESTAMPTZ NOT NULL DEFAULT now()
-    );
-      DO $$
-      BEGIN
-        IF NOT EXISTS (SELECT 1 FROM pg_constraint WHERE conname = 'billing_quotes_brand_fkey') THEN
-          ALTER TABLE billing_quotes ADD CONSTRAINT billing_quotes_brand_fkey FOREIGN KEY (brand) REFERENCES public.brands(id) ON UPDATE CASCADE ON DELETE RESTRICT NOT VALID;
-        END IF;
-      END $$;
-  `);
-  await pool.query(`
-    ALTER TABLE billing_invoices
-      ADD COLUMN IF NOT EXISTS hash_sha256    TEXT,
-      ADD COLUMN IF NOT EXISTS pdf_mime       TEXT,
-      ADD COLUMN IF NOT EXISTS pdf_size_bytes INTEGER,
-      ADD COLUMN IF NOT EXISTS finalized_at   TIMESTAMPTZ
-  `);
-  await pool.query(`ALTER TABLE billing_customers ADD COLUMN IF NOT EXISTS leitweg_id VARCHAR(46)`);
-  await pool.query(`CREATE INDEX IF NOT EXISTS idx_billing_customers_leitweg ON billing_customers(leitweg_id) WHERE leitweg_id IS NOT NULL`);
-  // Plan F: EU supply + export evidence
-  await pool.query(`
-    CREATE TABLE IF NOT EXISTS billing_nachweis (
-      id           BIGSERIAL PRIMARY KEY,
-      invoice_id   TEXT NOT NULL REFERENCES billing_invoices(id),
-      brand        TEXT REFERENCES public.brands(id) ON UPDATE CASCADE ON DELETE RESTRICT NOT NULL,
-      type         TEXT NOT NULL,
-      received_at  DATE,
-      document_ref TEXT,
-      notes        TEXT,
-      created_at   TIMESTAMPTZ NOT NULL DEFAULT now()
-    );
-      DO $$
-      BEGIN
-        IF NOT EXISTS (SELECT 1 FROM pg_constraint WHERE conname = 'billing_nachweis_brand_fkey') THEN
-          ALTER TABLE billing_nachweis ADD CONSTRAINT billing_nachweis_brand_fkey FOREIGN KEY (brand) REFERENCES public.brands(id) ON UPDATE CASCADE ON DELETE RESTRICT NOT VALID;
-        END IF;
-      END $$;
-  `);
-  // Plan F: VAT ID validation log
-  await pool.query(`
-    CREATE TABLE IF NOT EXISTS vat_id_validations (
-      id                  BIGSERIAL PRIMARY KEY,
-      customer_id         TEXT REFERENCES billing_customers(id),
-      vat_id              TEXT NOT NULL,
-      country_code        CHAR(2) NOT NULL,
-      valid               BOOLEAN NOT NULL,
-      vies_name           TEXT,
-      vies_address        TEXT,
-      request_identifier  TEXT,
-      validated_at        TIMESTAMPTZ NOT NULL DEFAULT now()
-    )
-  `);
-  await pool.query(`
-    CREATE TABLE IF NOT EXISTS billing_suppliers (
-      id            TEXT PRIMARY KEY DEFAULT gen_random_uuid()::text,
-      brand         TEXT REFERENCES public.brands(id) ON UPDATE CASCADE ON DELETE RESTRICT NOT NULL,
-      name          TEXT NOT NULL,
-      email         TEXT,
-      land_iso      CHAR(2) NOT NULL DEFAULT 'DE',
-      ustidnr       TEXT,
-      steuernummer  TEXT,
-      iban          TEXT,
-      bic           TEXT,
-      bank_name     TEXT,
-      address       TEXT,
-      typ           TEXT DEFAULT 'Lieferant',
-      created_at    TIMESTAMPTZ NOT NULL DEFAULT now(),
-      updated_at    TIMESTAMPTZ NOT NULL DEFAULT now(),
-      CONSTRAINT billing_suppliers_brand_name_key UNIQUE (brand, name)
-    );
-      DO $$
-      BEGIN
-        IF NOT EXISTS (SELECT 1 FROM pg_constraint WHERE conname = 'billing_suppliers_brand_fkey') THEN
-          ALTER TABLE billing_suppliers ADD CONSTRAINT billing_suppliers_brand_fkey FOREIGN KEY (brand) REFERENCES public.brands(id) ON UPDATE CASCADE ON DELETE RESTRICT NOT VALID;
-        END IF;
-      END $$;
-  `);
-  await pool.query(`
-    CREATE TABLE IF NOT EXISTS supplier_invoices (
-      id            TEXT PRIMARY KEY DEFAULT gen_random_uuid()::text,
-      brand         TEXT REFERENCES public.brands(id) ON UPDATE CASCADE ON DELETE RESTRICT NOT NULL,
-      supplier_id   TEXT NOT NULL REFERENCES billing_suppliers(id),
-      invoice_number TEXT,
-      invoice_date  DATE NOT NULL,
-      leistungsdatum DATE,
-      net_amount    NUMERIC(12,2) NOT NULL,
-      vat_amount    NUMERIC(12,2) NOT NULL DEFAULT 0,
-      gross_amount  NUMERIC(12,2) NOT NULL,
-      vat_rate      NUMERIC(5,2)  NOT NULL DEFAULT 0,
-      currency      CHAR(3) NOT NULL DEFAULT 'EUR',
-      description   TEXT,
-      pdf_path      TEXT,
-      status        TEXT NOT NULL DEFAULT 'open',
-      paid_at       DATE,
-      locked        BOOLEAN NOT NULL DEFAULT false,
-      created_at    TIMESTAMPTZ NOT NULL DEFAULT now(),
-      updated_at    TIMESTAMPTZ NOT NULL DEFAULT now()
-    );
-      DO $$
-      BEGIN
-        IF NOT EXISTS (SELECT 1 FROM pg_constraint WHERE conname = 'supplier_invoices_brand_fkey') THEN
-          ALTER TABLE supplier_invoices ADD CONSTRAINT supplier_invoices_brand_fkey FOREIGN KEY (brand) REFERENCES public.brands(id) ON UPDATE CASCADE ON DELETE RESTRICT NOT VALID;
-        END IF;
-      END $$;
-  `);
-  // Plan F: indexes for new child tables
-  await pool.query(`
-    CREATE INDEX IF NOT EXISTS billing_nachweis_invoice_idx
-      ON billing_nachweis (invoice_id)
-  `);
-  await pool.query(`
-    CREATE INDEX IF NOT EXISTS vat_id_validations_customer_idx
-      ON vat_id_validations (customer_id)
-      WHERE customer_id IS NOT NULL
-  `);
-  // Plan F: billing_invoice_payments — rate at payment time
-  await pool.query(`
-    ALTER TABLE billing_invoice_payments
-      ADD COLUMN IF NOT EXISTS payment_currency_rate NUMERIC(12,6)
-  `);
-
-  // Audit Phase 3 & 4: Billing Schema Cleanup
-  await pool.query(`
-    CREATE TABLE IF NOT EXISTS billing_invoice_documents (
-      invoice_id TEXT NOT NULL REFERENCES billing_invoices(id) ON DELETE CASCADE,
-      format     TEXT NOT NULL,
-      content    BYTEA NOT NULL,
-      PRIMARY KEY (invoice_id, format)
-    )
-  `);
-
-  // Migrate blobs to billing_invoice_documents
-  await pool.query(`
-    DO $$
-    BEGIN
-      IF EXISTS (SELECT 1 FROM information_schema.columns WHERE table_name='billing_invoices' AND column_name='pdf_blob') THEN
-        INSERT INTO billing_invoice_documents (invoice_id, format, content)
-        SELECT id, 'pdf', pdf_blob FROM billing_invoices WHERE pdf_blob IS NOT NULL
-        ON CONFLICT DO NOTHING;
-      END IF;
-      IF EXISTS (SELECT 1 FROM information_schema.columns WHERE table_name='billing_invoices' AND column_name='pdf_a3_blob') THEN
-        INSERT INTO billing_invoice_documents (invoice_id, format, content)
-        SELECT id, 'pdf-a3', pdf_a3_blob FROM billing_invoices WHERE pdf_a3_blob IS NOT NULL
-        ON CONFLICT DO NOTHING;
-      END IF;
-      IF EXISTS (SELECT 1 FROM information_schema.columns WHERE table_name='billing_invoices' AND column_name='zugferd_xml') THEN
-        INSERT INTO billing_invoice_documents (invoice_id, format, content)
-        SELECT id, 'zugferd', zugferd_xml::bytea FROM billing_invoices WHERE zugferd_xml IS NOT NULL
-        ON CONFLICT DO NOTHING;
-      END IF;
-      IF EXISTS (SELECT 1 FROM information_schema.columns WHERE table_name='billing_invoices' AND column_name='factur_x_xml') THEN
-        INSERT INTO billing_invoice_documents (invoice_id, format, content)
-        SELECT id, 'factur-x', factur_x_xml::bytea FROM billing_invoices WHERE factur_x_xml IS NOT NULL
-        ON CONFLICT DO NOTHING;
-      END IF;
-      IF EXISTS (SELECT 1 FROM information_schema.columns WHERE table_name='billing_invoices' AND column_name='xrechnung_xml') THEN
-        INSERT INTO billing_invoice_documents (invoice_id, format, content)
-        SELECT id, 'xrechnung', xrechnung_xml::bytea FROM billing_invoices WHERE xrechnung_xml IS NOT NULL
-        ON CONFLICT DO NOTHING;
-      END IF;
-    END $$;
-  `);
-
-  // Drop redundant columns
-  await pool.query(`
-    DO $$
-    BEGIN
-      IF EXISTS (SELECT 1 FROM information_schema.columns WHERE table_name='billing_invoices' AND column_name='pdf_blob') THEN
-        ALTER TABLE billing_invoices DROP COLUMN pdf_blob;
-      END IF;
-      IF EXISTS (SELECT 1 FROM information_schema.columns WHERE table_name='billing_invoices' AND column_name='pdf_a3_blob') THEN
-        ALTER TABLE billing_invoices DROP COLUMN pdf_a3_blob;
-      END IF;
-      IF EXISTS (SELECT 1 FROM information_schema.columns WHERE table_name='billing_invoices' AND column_name='zugferd_xml') THEN
-        ALTER TABLE billing_invoices DROP COLUMN zugferd_xml;
-      END IF;
-      IF EXISTS (SELECT 1 FROM information_schema.columns WHERE table_name='billing_invoices' AND column_name='factur_x_xml') THEN
-        ALTER TABLE billing_invoices DROP COLUMN factur_x_xml;
-      END IF;
-      IF EXISTS (SELECT 1 FROM information_schema.columns WHERE table_name='billing_invoices' AND column_name='xrechnung_xml') THEN
-        ALTER TABLE billing_invoices DROP COLUMN xrechnung_xml;
-      END IF;
-      IF EXISTS (SELECT 1 FROM information_schema.columns WHERE table_name='billing_invoices' AND column_name='paid_at') THEN
-        ALTER TABLE billing_invoices DROP COLUMN paid_at;
-      END IF;
-      IF EXISTS (SELECT 1 FROM information_schema.columns WHERE table_name='billing_invoices' AND column_name='paid_amount') THEN
-        ALTER TABLE billing_invoices DROP COLUMN paid_amount;
-      END IF;
-      IF EXISTS (SELECT 1 FROM information_schema.columns WHERE table_name='billing_invoices' AND column_name='dunning_level') THEN
-        ALTER TABLE billing_invoices DROP COLUMN dunning_level;
-      END IF;
-      IF EXISTS (SELECT 1 FROM information_schema.columns WHERE table_name='billing_invoices' AND column_name='last_dunning_at') THEN
-        ALTER TABLE billing_invoices DROP COLUMN last_dunning_at;
-      END IF;
-    END $$;
-  `);
-
-  // Create view
-  await pool.query(`
-    CREATE OR REPLACE VIEW v_billing_invoices_with_state AS
-    SELECT
-      i.*,
-      COALESCE(p.paid_amount, 0) AS paid_amount,
-      CASE WHEN COALESCE(p.paid_amount, 0) >= i.gross_amount THEN p.last_paid_at ELSE NULL END AS paid_at,
-      COALESCE(d.dunning_level, 0) AS dunning_level,
-      d.last_dunning_at
-    FROM billing_invoices i
-    LEFT JOIN (
-      SELECT
-        invoice_id,
-        SUM(amount) AS paid_amount,
-        MAX(paid_at) AS last_paid_at
-      FROM billing_invoice_payments
-      GROUP BY invoice_id
-    ) p ON i.id = p.invoice_id
-    LEFT JOIN (
-      SELECT
-        invoice_id,
-        MAX(level) AS dunning_level,
-        MAX(generated_at) AS last_dunning_at
-      FROM billing_invoice_dunnings
-      GROUP BY invoice_id
-    ) d ON i.id = d.invoice_id;
-  `);
-
-  await initBillingAuditTable();
-  await installInvoiceImmutabilityTriggers();
-  billingTablesReady = true;
-}
-
-let taxModeTableReady = false;
-export async function initTaxMonitorTables(): Promise<void> {
-  if (taxModeTableReady) return;
-  await pool.query(`
-    CREATE TABLE IF NOT EXISTS tax_mode_changes (
-      id            BIGSERIAL PRIMARY KEY,
-      brand         TEXT REFERENCES public.brands(id) ON UPDATE CASCADE ON DELETE RESTRICT NOT NULL,
-      changed_at    TIMESTAMPTZ NOT NULL DEFAULT now(),
-      from_mode     TEXT NOT NULL,
-      to_mode       TEXT NOT NULL,
-      trigger_invoice_id TEXT,
-      year_revenue_at_change NUMERIC(12,2),
-      notes         TEXT
-    );
-      DO $$
-      BEGIN
-        IF NOT EXISTS (SELECT 1 FROM pg_constraint WHERE conname = 'tax_mode_changes_brand_fkey') THEN
-          ALTER TABLE tax_mode_changes ADD CONSTRAINT tax_mode_changes_brand_fkey FOREIGN KEY (brand) REFERENCES public.brands(id) ON UPDATE CASCADE ON DELETE RESTRICT NOT VALID;
-        END IF;
-      END $$;
-  `);
-  taxModeTableReady = true;
-}
-
-let eurTablesReady = false;
-export async function initEurTables(): Promise<void> {
-  if (eurTablesReady) return;
-  await pool.query(`
-    CREATE TABLE IF NOT EXISTS eur_bookings (
-      id            BIGSERIAL PRIMARY KEY,
-      brand         TEXT REFERENCES public.brands(id) ON UPDATE CASCADE ON DELETE RESTRICT NOT NULL,
-      booking_date  DATE NOT NULL,
-      type          TEXT NOT NULL,
-      category      TEXT NOT NULL,
-      description   TEXT NOT NULL,
-      net_amount    NUMERIC(12,2) NOT NULL,
-      vat_amount    NUMERIC(12,2) NOT NULL DEFAULT 0,
-      invoice_id    TEXT REFERENCES billing_invoices(id),
-      receipt_path  TEXT,
-      created_at    TIMESTAMPTZ NOT NULL DEFAULT now()
-    );
-      DO $$
-      BEGIN
-        IF NOT EXISTS (SELECT 1 FROM pg_constraint WHERE conname = 'eur_bookings_brand_fkey') THEN
-          ALTER TABLE eur_bookings ADD CONSTRAINT eur_bookings_brand_fkey FOREIGN KEY (brand) REFERENCES public.brands(id) ON UPDATE CASCADE ON DELETE RESTRICT NOT VALID;
-        END IF;
-      END $$;
-  `);
-  await pool.query(`
-    ALTER TABLE eur_bookings
-      ADD COLUMN IF NOT EXISTS belegnummer TEXT,
-      ADD COLUMN IF NOT EXISTS skr_konto   TEXT
-  `);
-  await pool.query(`
-    CREATE TABLE IF NOT EXISTS assets (
-      id                   BIGSERIAL PRIMARY KEY,
-      brand                TEXT REFERENCES public.brands(id) ON UPDATE CASCADE ON DELETE RESTRICT NOT NULL,
-      description          TEXT NOT NULL,
-      purchase_date        DATE NOT NULL,
-      net_purchase_price   NUMERIC(12,2) NOT NULL,
-      vat_paid             NUMERIC(12,2) NOT NULL,
-      useful_life_months   INT NOT NULL,
-      correction_start_date DATE,
-      is_gwg               BOOLEAN NOT NULL DEFAULT false,
-      receipt_path         TEXT,
-      created_at           TIMESTAMPTZ NOT NULL DEFAULT now()
-    );
-      DO $$
-      BEGIN
-        IF NOT EXISTS (SELECT 1 FROM pg_constraint WHERE conname = 'assets_brand_fkey') THEN
-          ALTER TABLE assets ADD CONSTRAINT assets_brand_fkey FOREIGN KEY (brand) REFERENCES public.brands(id) ON UPDATE CASCADE ON DELETE RESTRICT NOT VALID;
-        END IF;
-      END $$;
-  `);
-  eurTablesReady = true;
-}
 
 // ─── Service-page content store (per-slug) ────────────────────────────────────
